fix: build by moving test things out of @services/logseq
You cannot import test files in module export
