+++ src/queries/getBlockBy.spec.ts
@@ -1,7 +1,9 @@
-import { generateMoqseqClient } from "src/services/logseq/mock/client.js";
+import { TESTING_FUNCS } from "@testing/logseq";
 import { describe, expect, it } from "vitest";
 import { findPagesByRaindropID } from "./getBlockBy.js";
 import type { LSBlockEntity, LSPageEntity } from "@services/logseq";
+
+const { generateMoqseqClient } = TESTING_FUNCS;
 
 // Created by running the query "(property raindrop-id 457389317)" on
 // my Logseq database, and then removing the following non-supported fields:
