+++ src/services/logseq/_testing/index.ts
@@ -1,0 +1,9 @@
+import type { PageEntityWithRootBlocks } from "./mock/types.js";
+import { generateMoqseqClient } from "./mock/client.js";
+
+export type TESTING_TYPES = {
+  PageEntityWithRootBlocks: PageEntityWithRootBlocks;
+};
+export const TESTING_FUNCS = {
+  generateMoqseqClient,
+};
