+++ src/services/logseq/_testing/mock/types.ts
@@ -1,5 +1,5 @@
 import type { BlockUUID } from "@logseq/libs/dist/LSPlugin.js";
-import type { LSBlockEntity, LSPageEntity } from "../interfaces.js";
+import type { LSBlockEntity, LSPageEntity } from "../../interfaces.js";
 
 export type PageEntityWithRootBlocks = LSPageEntity & {
   roots?: ["uuid", BlockUUID][];
