+++ src/services/logseq/index.ts
@@ -15,14 +15,3 @@ export {
   importFilterOptions,
   registerSettings,
 } from "./settings.js";
-
-// Used for testing
-import type { PageEntityWithRootBlocks } from "./mock/types.js";
-import { generateMoqseqClient } from "./mock/client.js";
-
-export type TESTING_TYPES = {
-  PageEntityWithRootBlocks: PageEntityWithRootBlocks;
-};
-export const TESTING_FUNCS = {
-  generateMoqseqClient,
-};
